# Remove deprecated aho corasick `.byte_classes(false)` toggle